feat: add command and query bus
